changed tree processor to repo detector
